+++ gpu/srcs/server/master.c
@@ -61,8 +61,8 @@ static void		send_offsets(t_client *clients, t_cluster *cluster, int nbr)
 	{
 		if (clients)
 		{
-			offsets.y_min = nbr * WIN_HEIGHT / cluster->nbr_clients;
-			offsets.y_max = offsets.y_min + WIN_HEIGHT / cluster->nbr_clients;
+			offsets.y_min = nbr * (double)WIN_HEIGHT / cluster->nbr_clients;
+			offsets.y_max = offsets.y_min + (double)WIN_HEIGHT / cluster->nbr_clients;
 			offsets.render_factor = cluster->world->render_factor;
 			offsets.aa = cluster->world->aa;
 			offsets.ob_save = cluster->world->ob_save;
